refactor: Hilt 의존성 주입이 무조건SingletonComponent로 선언되어 있는 문제 해결
